Fix google sign in error. (#48)
Because:

When users would try the demo, we weren't accounting for there being
a user in the session that also signs up via google. The `if`
statement would short circuit when it found a `current_user` and ignore
the `google_id_token`.

This:

Extracts a `Registration` form object that cleans up the logic and
fixes the issue above.