Allow users to continue planning tomorrow.
Because:

When a user completes their reflection and taken to the tomorrow path,
they may navigate to the dashboard. If they do this, they won't be able
to get back to the `/tomorrow` route.

This:

Makes the button on `today#show` dynamically link to the `/tomorrow`
path if they have already written their reflection.